+++ tests/MiddlewareAndComponentTest.php
@@ -45,7 +45,7 @@ class MiddlewareAndComponentTest extends BaseTest
         Http::assertSent(function ($request) {
             return $request['blocks'][0] === [
                 'id' => 'component',
-                'hash' => 'e99681f5450cbaf3774adc5eb74d637f',
+                'hash' => 'fe47ca550596fd0cea4664cc73774bcb',
                 'language' => 'php',
                 'theme' => 'material-theme-palenight',
                 'code' => 'echo "hello world";',